+++ src/main/java/fr/inrae/act/bagap/chloe/counting/ValueCountingInterface.java
@@ -6,19 +6,19 @@ public interface ValueCountingInterface {
 	 * to get values of the map
 	 * @return values of the map
 	 */
-	short[] values();
+	int[] values();
 	
 	/**
 	 * to count value of type v
 	 * @param v : type of value
 	 * @return the count of v
 	 */
-	double countValue(short v);
+	double countValue(int v);
 	
 	/**
 	 * to count the classes
 	 * @return the count of classes
 	 */
-	short countClass();
+	int countClass();
 	
 }
