+++ src/main/java/fr/inrae/act/bagap/chloe/metric/value/ShannonDiversityIndex.java
@@ -14,7 +14,7 @@ public class ShannonDiversityIndex extends Metric implements ValueMetric {
 		if(co.countValues() > 0){
 			value = 0;
 			double p; 
-			for(short v : co.values()){
+			for(int v : co.values()){
 				//System.out.println(v+" "+co.countValue(v));
 				p = co.countValue(v) / co.validValues();
 				if(p != 0){
